refactor: criado useProfessional hook
