+++ TODO.md
@@ -5,7 +5,7 @@
 [ ] Integração Roles [x] Create [x] Read [x] Show [x] Update [x] [ ] Delete
 [ ] Integração Services [x] Create [x] Read [x] Show [x] Update [x] [ ] Delete
 [ ] Integração Patients [x] Create [x] Read [x] Show [x] Update [x] [ ] Delete
-[ ] Integração Professionals [ ] Create [ ] Read [ ] Show [ ] Update [ ] [ ] Delete
+[ ] Integração Professionals [x] Create [x] Read [ ] Show [ ] Update [ ] [ ] Delete
 [ ] Integração Appointments[ ] Create [ ] Read [ ] Show [ ] Update [ ] [ ] Delete
 
 # OUTROS
